fix: navbar, slider 카테고리 컴포넌트 분리, footer 수정
