+++ src/components/Footer/Footer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import { notice, bankAccount, footerTop } from './footerData';
+import { notice, bankAccount, footerTop, csCenterNumber } from './footerData';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faMicrophoneAlt } from '@fortawesome/free-solid-svg-icons';
 import { faCommentAlt } from '@fortawesome/free-regular-svg-icons';
@@ -46,7 +46,7 @@ class Footer extends Component {
               <div className="line">_______</div>
               <div className="csCenterNumber">
                 <FontAwesomeIcon icon={faMicrophoneAlt} />
-                <h2>&nbsp;8282-8282</h2>
+                <h2>&nbsp;{csCenterNumber}</h2>
               </div>
               <p>
                 매일 오후12시 ~ 오후11시59분
