+++ src/components/Footer/footerData.js
@@ -2,12 +2,14 @@ const notice = [
   { link: '/notice/1', name: '10월에도 행복하세요' },
   { link: '/notice/2', name: '9월에 참 행복했죠?' },
 ];
+
 const bankAccount = [
   { name: '다빈은행', account: '123412341234' },
   { name: '지현은행', account: '123412341234' },
   { name: '정호은행', account: '123412341234' },
   { name: '재원은행', account: '123412341234' },
 ];
+
 const footerTop = [
   { link: '/privacy', name: '개인정보처리방침' },
   { link: '/cooperate', name: '제휴/입점문의' },
@@ -15,4 +17,6 @@ const footerTop = [
   { link: '/cscenter', name: '자주하는 질문' },
 ];
 
-export { notice, bankAccount, footerTop };
+const csCenterNumber = '8282-8282';
+
+export { notice, bankAccount, footerTop, csCenterNumber };
