+++ src/components/NavBar/NavBar.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
 import { navbarList } from './category';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faGripLines, faBox } from '@fortawesome/free-solid-svg-icons';
+import { faGripLines } from '@fortawesome/free-solid-svg-icons';
+import MainCategory from '../MainCategory/MainCategory';
+import NavBarList from './NavBarList/NavBarList';
 import './NavBar.scss';
 
 class NavBar extends Component {
@@ -23,10 +24,6 @@ class NavBar extends Component {
       );
   }
 
-  moveToCategory = link => {
-    window.location.replace(`/category/${link}`);
-  };
-
   render() {
     const { category } = this.state;
     return (
@@ -40,39 +37,13 @@ class NavBar extends Component {
             <div className="dropDownCategory">
               <ul className="mainMenu">
                 {category.map((data, id) => {
-                  const name = '/list/' + data.id;
-                  return (
-                    <li className="mainMenuList" key={id}>
-                      <Link to={name}>{data.name}</Link>
-                      <ul className="subMenu">
-                        {data.list.map((list, id) => {
-                          const name = '/list/' + data.id + '/' + list.id;
-                          return (
-                            <li key={id}>
-                              <Link to={name}>{list.name}</Link>
-                            </li>
-                          );
-                        })}
-                      </ul>
-                    </li>
-                  );
+                  return <MainCategory key={id} data={data} />;
                 })}
               </ul>
             </div>
           </li>
           {navbarList.map((list, id) => {
-            return (
-              <li
-                className="navbarList"
-                key={id}
-                onClick={() => {
-                  this.moveToCategory(list.link);
-                }}
-              >
-                {list.name.includes('배송') && <FontAwesomeIcon icon={faBox} />}
-                &nbsp;{list.name}
-              </li>
-            );
+            return <NavBarList key={id} list={list} />;
           })}
         </ul>
       </nav>
